POLIO-2067: add TS typing
