+++ plugins/polio/js/src/domains/Calendar/campaignCalendar/index.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react';
 import { Box, Table, TableContainer } from '@mui/material';
 import { LoadingSpinner } from 'bluesquare-components';
 import { Moment } from 'moment';
-import { CalendarOrdering } from '../hooks/useMergedCampaigns';
+import { CalendarOrdering } from '../hooks/useMergedCampaigns/useMergedCampaigns';
 import { Body } from './Body';
 import { Head } from './Head';
 import { Nav } from './Nav';
