+++ plugins/polio/js/src/domains/Calendar/campaignCalendar/utils/index.tsx
@@ -1,7 +1,9 @@
 import React, { ReactElement } from 'react';
+import { SxProps } from '@mui/material';
 import { Moment } from 'moment';
 
 import { EmptyCell } from '../cells/Empty';
+import { INTEGRATED_CAMPAIGN_BORDER_COLOR } from '../constants';
 import { MappedCampaign } from '../types';
 
 export const isStartOfWeek = (date: Moment): boolean => date.isoWeekday() === 1; // 1 is Monday
@@ -222,4 +224,68 @@ export const addRemainingEmptyCells = ({
         currentWeekIndex,
         id,
     });
+};
+export const getStaticFieldCellStyle = (
+    fieldKey: string,
+    campaign: MappedCampaign,
+): SxProps => {
+    const borderStyle = `2px solid ${INTEGRATED_CAMPAIGN_BORDER_COLOR}`;
+    if (fieldKey === 'edit')
+        switch (campaign.layout) {
+            case 'top':
+                return {
+                    borderTop: borderStyle,
+                    borderLeft: borderStyle,
+                };
+            case 'middle':
+                return {
+                    borderLeft: borderStyle,
+                };
+            case 'bottom':
+                return {
+                    borderBottom: borderStyle,
+                    borderLeft: borderStyle,
+                };
+            default:
+                return {};
+        }
+    if (
+        fieldKey === 'country' ||
+        fieldKey == 'name' ||
+        fieldKey === 'campaign_types'
+    ) {
+        switch (campaign.layout) {
+            case 'top':
+                return {
+                    borderTop: borderStyle,
+                };
+            case 'middle':
+                return {};
+            case 'bottom':
+                return {
+                    borderBottom: borderStyle,
+                };
+            default:
+                return {};
+        }
+    }
+    if (fieldKey === 'r1StartDate') {
+        switch (campaign.layout) {
+            case 'top':
+                return {
+                    borderTop: borderStyle,
+                    borderRight: borderStyle,
+                };
+            case 'middle':
+                return { borderRight: borderStyle };
+            case 'bottom':
+                return {
+                    borderBottom: borderStyle,
+                    borderRight: borderStyle,
+                };
+            default:
+                return {};
+        }
+    }
+    return {};
 };
